Migrate pipeline.py to repype
